+++ manuals/sources/conf.py
@@ -28,7 +28,7 @@ author = u'Paulo Moura'
 # The short X.Y version
 version = '3.39.0'
 # The full version, including alpha/beta/rc tags
-release = u'v3.39.0-b03'
+release = u'v3.39.0-b04'
 
 
 # -- General configuration ---------------------------------------------------